+++ baselib/src/main/java/com/salieri/baselib/utils/BaseUtil.java
@@ -2,7 +2,7 @@ package com.salieri.baselib.utils;
 
 
 public class BaseUtil {
-    public static double degree2Radius (float degree) {
-        return (degree / 360) * 2 * Math.PI;
+    public static double degree2Radius (double degree) {
+        return (degree / 360d) * 2d * Math.PI;
     }
 }
